requirements.txt
django==5.0.7
django-bootstrap5
django-ckeditor==6.2.0
django-phonenumber-field[phonenumbers]
django-localflavor
django-modeltranslation
Pillow
gunicorn
whitenoise
django-heroku
Faker
django-nested-admin
django-cms~=4.1
djangocms-alias
django-sekizai
djangocms-text-ckeditor
djangocms-frontend==1.3.5
celery
redis
selenium
django-celery-beat
django-celery-results
joeflow[reversion,celery]
django-simple-history
packaging
sib-api-v3-sdk
beautifulsoup4
html-sanitizer
requests
django-crispy-forms
crispy-bootstrap5
splinter[django]
django-anymail[sendinblue,mailjet]
supabase
slack_sdk
django-htmx
openai
presidio-analyzer==2.2.354
presidio-anonymizer==2.2.354
en-core-web-lg @ https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.7.1/en_core_web_lg-3.7.1-py3-none-any.whl
